Organise tests into separate folders.
