+++ lessons/tests/views/test_home_view.py
@@ -1,8 +1,7 @@
-from django.test import SimpleTestCase
+from django.test import TestCase
 from django.urls import reverse
-from django.http import HttpResponse
 
-class HomePageTest(SimpleTestCase):
+class HomeViewTestCase(TestCase):
     def test_url_exists_at_correct_location(self):
         response = self.client.get("/")
         self.assertEqual(response.status_code, 200)
